[Fix/BE] user id 조회 수정

+++ backend/auth-server/src/main/java/instagram_clone/sgdevcamp_jikji_insta_clone_auth_server/user/repository/UserRepository.java
@@ -15,6 +15,6 @@ public interface UserRepository extends JpaRepository<User, String>{
 
 	List<User> findAll();
 
-	Optional<User> findById(Long id);
+	Optional<User> findById(Integer id);
 }
